Now with memcache support : erl -boot start_sasl -s s3 start -sname test -s3 memcached '{"localhost", 11211}'
(with of course memcached running on localhost:11211), and merle installed (http://github.com/joewilliams/merle/tree/master)
Only objects are stored.
Be careful, it's currently using merle and terms. It is not expected to be portable with other languages.
Next step is ketama'izing the lookup for accessing multiple memcached servers.